aws/cloud-development-kit/Stack/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws_s3_deployment
aws_cdk.aws_dynamodb
aws_cdk.aws_events
aws_cdk.aws_lambda
aws_cdk.aws_events_targets